backend/requirements.txt
# FastAPI Backend
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
websockets>=12.0

# LangChain & AI
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
langchain-huggingface>=0.1.0
langchain-text-splitters>=0.3.0

# Embeddings & Vector Store
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0

# PDF Processing
pdfplumber>=0.11.0

# Utilities
torch>=2.0.0
pydantic>=2.0.0
